Move details feedback form component into sub directory
